requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.72
botocore==1.31.72
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.1.13
django-cors-headers==4.1.0
django-filter==23.2
django-otp==1.2.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
filetype==1.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
Pillow==9.5.0
psycopg2==2.9.7
PyJWT==2.7.0
pypng==0.20220715.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.8.9
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
yarl==1.9.2
